+++ part2/countries/src/components/CountryEntry.js
@@ -1,0 +1,26 @@
+import { useState } from "react"
+import Countryview from "./CountryView"
+
+const Countryentry = ({country}) => {
+    const [displayCountry, setDisplayCountry] = useState(false)
+
+    const handleClick = () => {
+        setDisplayCountry(true)
+    }
+
+    if (!displayCountry) {
+        return (
+            <div>
+                <span>{country.name["official"]}</span> <input type="submit" value="show" onClick={handleClick}></input>
+            </div>
+        )
+    }
+    else {
+        return (
+            <Countryview country={country}/>
+        )
+    }
+
+}
+
+export default Countryentry
